+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Settings screen
+- Logout button
+- Bottom app bar
 
 ## [1.0.0] - 2021-06-18
 ### Added